Rename old dashboard folder
